# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Quiz Platform API",
        "description": "API documentation for Quiz Platform mobile app",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/admin/questions": {
            "get": {
                "tags": [
                    "Admin Questions"
                ],
                "description": "List questions with filtering.",
                "operationId": "ddad86441ec7650058d36ddf17fa6018",
                "parameters": [
                    {
                        "name": "category",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "difficulty",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of questions"
                    },
                    "403": {
                        "description": "Forbidden - Admin only"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin Questions"
                ],
                "description": "Create a new question.",
                "operationId": "adc662728a8214706967f5588bfa0074",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "category_id": {
                                        "description": "Category ID",
                                        "type": "integer"
                                    },
                                    "question_text": {
                                        "description": "Question text",
                                        "type": "string"
                                    },
                                    "question_text_ar": {
                                        "description": "Question text (Arabic)",
                                        "type": "string"
                                    },
                                    "answer_a": {
                                        "description": "Answer A",
                                        "type": "string"
                                    },
                                    "answer_b": {
                                        "description": "Answer B",
                                        "type": "string"
                                    },
                                    "answer_c": {
                                        "description": "Answer C",
                                        "type": "string"
                                    },
                                    "answer_d": {
                                        "description": "Answer D",
                                        "type": "string"
                                    },
                                    "correct_answer": {
                                        "description": "Correct answer letter",
                                        "type": "string",
                                        "enum": [
                                            "a",
                                            "b",
                                            "c",
                                            "d"
                                        ]
                                    },
                                    "difficulty": {
                                        "description": "Difficulty level",
                                        "type": "string",
                                        "enum": [
                                            "medium",
                                            "hard",
                                            "very_hard"
                                        ]
                                    },
                                    "status": {
                                        "description": "Question status",
                                        "type": "string",
                                        "enum": [
                                            "active",
                                            "inactive"
                                        ]
                                    },
                                    "image": {
                                        "description": "Image (max 2MB)",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Question created"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/admin/questions/{question}": {
            "get": {
                "tags": [
                    "Admin Questions"
                ],
                "description": "Get a single question details.",
                "operationId": "7d3fdf49ca4c94e3b79b12645005190a",
                "parameters": [
                    {
                        "name": "question",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Question details"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin Questions"
                ],
                "description": "Update a question.",
                "operationId": "d45bcef7f84450d3d1ee4b8d17b96979",
                "parameters": [
                    {
                        "name": "question",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "category_id": {
                                        "type": "integer"
                                    },
                                    "question_text": {
                                        "type": "string"
                                    },
                                    "question_text_ar": {
                                        "type": "string"
                                    },
                                    "answer_a": {
                                        "type": "string"
                                    },
                                    "answer_b": {
                                        "type": "string"
                                    },
                                    "answer_c": {
                                        "type": "string"
                                    },
                                    "answer_d": {
                                        "type": "string"
                                    },
                                    "correct_answer": {
                                        "type": "string",
                                        "enum": [
                                            "a",
                                            "b",
                                            "c",
                                            "d"
                                        ]
                                    },
                                    "difficulty": {
                                        "type": "string",
                                        "enum": [
                                            "medium",
                                            "hard",
                                            "very_hard"
                                        ]
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "active",
                                            "inactive"
                                        ]
                                    },
                                    "image": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Question updated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Admin Questions"
                ],
                "description": "Delete a question.",
                "operationId": "ab499abba7ed6ed90739255e8e31c519",
                "parameters": [
                    {
                        "name": "question",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Question deleted"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/admin/questions-submissions": {
            "get": {
                "tags": [
                    "Admin Questions"
                ],
                "description": "List question submissions.",
                "operationId": "117c81bda718fed6f5fe07281eb8c5cf",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter by status (pending, approved, rejected)",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of submissions"
                    },
                    "403": {
                        "description": "Forbidden - Admin only"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/admin/questions-submissions/{submission}": {
            "get": {
                "tags": [
                    "Admin Questions"
                ],
                "description": "Get a single submission for review.",
                "operationId": "53e4258ff6357537a6bbf762ff663001",
                "parameters": [
                    {
                        "name": "submission",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Submission details"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/admin/questions-submissions/{submission}/approve": {
            "post": {
                "tags": [
                    "Admin Questions"
                ],
                "description": "Approve a submitted question.",
                "operationId": "a150fdf0be74953f76b57bd7b9beae74",
                "parameters": [
                    {
                        "name": "submission",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Submission approved"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/admin/questions-submissions/{submission}/reject": {
            "post": {
                "tags": [
                    "Admin Questions"
                ],
                "description": "Reject a submitted question with a reason.",
                "operationId": "ecafb9f2e27f326d53a569f403d17522",
                "parameters": [
                    {
                        "name": "submission",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "reason": {
                                        "description": "Rejection reason (max 500 chars)",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Submission rejected"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearer_token": {
                "type": "http",
                "description": "Enter your bearer token",
                "scheme": "bearer"
            },
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}